ml/Nutrition Science RAG/requirements.txt
flask==3.0.0
anthropic==0.3.11
pandas==2.2.0
openpyxl==3.1.2
langchain==0.1.13
langchain-community==0.0.31
langchain-text-splitters==0.0.5
langchain-huggingface==0.0.8
faiss-cpu==1.7.4
python-dotenv==1.0.1
sentence-transformers==2.5.1
